server/requirements.txt
Flask
Flask-CORS
python-dotenv
google-genai